p4i integration Integrate 179427 from 4.0 to main:        another "shuttup" fix
[git-p4: depot-paths = "//depot/qt/main/": change = 180934]

+++ src/gui/painting/qblackraster.c
@@ -1177,7 +1177,7 @@ Bezier_Up( RAS_ARGS Int        degree,
 Fin:
     ras.top  = top;
     ras.arc -= degree;
-    QT_FT_TRACE6("    currently %ld points in profile, start=%ld top=%p\n", ras.top - ras.cProfile->offset, ras.cProfile->start, ras.top);
+    QT_FT_TRACE6("    currently %d points in profile, start=%ld top=%p\n", (int)(ras.top - ras.cProfile->offset), ras.cProfile->start, ras.top);
     return SUCCESS;
 }
